+++ demo_plugins/聊天栏菜单/__init__.py
@@ -1,7 +1,9 @@
-from tooldelta import plugins, Plugin, PluginAPI, Frame, Builtins, Config
+from tooldelta import plugins, Plugin, PluginAPI
 
 from dataclasses import dataclass
 from typing import Callable
+
+from tooldelta.Frame import Builtins, Config, Frame
 
 plugins.checkSystemVersion((0, 1, 8))
